Fixed indents in data tracker.css
Looked in Prod and some of the CSS had double indents previously. Not sure if that's why when copy/paste in IDE/Notion lead to removed indents.

I removed the double indents which appears to fix the indent issue when copy/paste data tracker to IDE.

Also fixed the `overflow-wrap: break word` on line 273.